Fix types for providers
