+++ frontend/src/metabase/auth/selectors.ts
@@ -1,18 +1,15 @@
 import { createSelector } from "@reduxjs/toolkit";
-
 import { getSettings } from "metabase/selectors/settings";
 import { PLUGIN_AUTH_PROVIDERS } from "metabase/plugins";
+import { AuthProvider } from "metabase/plugins/types";
 
-import type { AuthProvider } from "./types";
+const EMPTY_PROVIDERS: AuthProvider[] = [];
 
-export const getAuthProviders = createSelector(
-  [getSettings],
-  (): AuthProvider[] =>
-    PLUGIN_AUTH_PROVIDERS.reduce(
-      (providers: any, getProviders: (providers: any) => any) =>
-        getProviders(providers),
-      [],
-    ),
+export const getAuthProviders = createSelector([getSettings], () =>
+  PLUGIN_AUTH_PROVIDERS.reduce(
+    (providers, getProviders) => getProviders(providers),
+    EMPTY_PROVIDERS,
+  ),
 );
 
 export const getExternalAuthProviders = createSelector(
