+++ frontend/src/metabase/auth/types.ts
@@ -1,20 +1,3 @@
-import { ComponentType } from "react";
-
-export interface AuthProvider {
-  name: string;
-  Button: ComponentType<AuthProviderButtonProps>;
-  Panel?: ComponentType<AuthProviderPanelProps>;
-}
-
-export interface AuthProviderButtonProps {
-  isCard?: boolean;
-  redirectUrl?: string;
-}
-
-export interface AuthProviderPanelProps {
-  redirectUrl?: string;
-}
-
 export interface LoginData {
   username: string;
   password: string;
